// src/server/model/iItemGroup.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroup } from './iAttributeGroup';

export interface IItemGroup { 
    parentItemGroupId?: number;
    parentItemGroupName?: string;
    systemGenerated?: boolean;
    filterAttributeGroupName?: string;
    filterAttributeGroupId?: number;
    featureAttributeGroups?: Array<IAttributeGroup>;
    hsnCode?: string;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}